[EL-84] PUSH-уведомления. Часть 1

+++ feature/notifications/implementation/src/main/java/ru/dekabrsky/feature/notifications/implementation/domain/entity/NotificationEntity.kt
@@ -1,12 +1,8 @@
 package ru.dekabrsky.feature.notifications.implementation.domain.entity
 
-import androidx.room.PrimaryKey
-import org.threeten.bp.LocalDate
-import java.util.Date
-
 @Suppress("LongParameterList")
 class NotificationEntity(
-    val uid: Int? = null,
+    val uid: Long? = null,
     val tabletName: String = "",
     val dosage: String = "",
     val note: String = "",
